requirements.txt
# ============================================================
# PolyBot - Production Dependencies
# ============================================================
# Install: pip install -r requirements.txt
# ============================================================

# ── Web Framework ──────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0

# ── HTTP & WebSocket ──────────────────────────────────────
requests>=2.31.0
websocket-client>=1.7.0

# ── Environment & Config ──────────────────────────────────
python-dotenv>=1.0.0

# ── Polymarket SDKs ───────────────────────────────────────
py-clob-client>=0.34.0
poly-web3>=1.0.0

# ── Blockchain & Crypto ───────────────────────────────────
web3>=7.0.0
eth-abi>=5.0.0
eth-utils>=5.0.0
